Update PolicyLambdaProvision for skunkworks
Throughout the test suite we use fixtures recorded against an old AWS
account that is no longer in use on the project. Over in #869, we want
to add a test to the PolicyLambdaProvision test class that uses
self.role, but first we need to update to a role in the new skunkworks
AWS account. Moreover, since we never run CI against live AWS, we have
some additional drift to account for here.