Localize import hack in custodian_archive
We want to keep mu decoupled from custodian as much as we can. This
commit replaces the virtualenv_dir parameter of PythonPackageArchive
with lib_paths, since the former was only used for computing
venv_lib_paths anyway. Effectively then we are pushing the
responsibility for computing lib_paths up to the caller.