Adding Region field to the request limit increase
I have added a region field to the body of the request limit increase email as many AWS technicians miss the region in the subject and then reply to my auto-generated cases asking me to confirm what region to increase the limit in.  Adding it to the case body will make this a lot clearer and prevent users from having to manually respond to AWS to complete the limit increases.