Add a S3 has-policy-permission filter
This filter looks for specific S3 permissions (actions and principals)
in a S3 bucket policy.  It works even when bucket policies have
wildcards in actions.

For example, it can be hard to determine if a bucket policy grants write
access with existing filters (such as `has-statement`) because actions
can have wildcards.  For example, any of the following actions would
grant write access to a bucket:

* s3:*
* s3:Put*
* s3:PutObject

By using the `has-policy-permission` filter, users can specify the
permission(s) they are looking for and this filter will match if the
bucket policy grants that permission.