Adding a Lambda version, removing filters, and changing condition
I added a lambda version of this policy to set the policy on all future buckets.  I found I had to change the condition from StringEquals to StringEqualsIgnoreCaseIfExists for this to block the "Make Public" button in the console.
I removed the filters as well as they didn't work the greatest as customers could modify the bucket policy after the fact and it wouldn't always get picked up on as being changed.