Create assoc/unassoc iters directly, bye partition
Rather than have a partition function to split the list of
network addresses into associated and unassociated for appropriate
handling, just create the generators explicitly (and inline) but
retain a function (lambda) for the split logic and have the 2
generators created differ only by the negation.

Goal all along is to make clearer the logic between how each
generator is created so that hopefully it is clear there is a
single logic function that needs to be updated to change the split
whereas copying the actual logic makes it less clear that the
behaviors between the 2 generators must be exact complements.