fix: Fixing the S3 bucket policy custodian applies
The S3 policy as it was would block the AWS Log Delivery service from writing logs to any buckets the policy was applied to.  By changing the Principle: * to a NotPrinciple: AWS: "arn:aws:iam::858827067514:root" that fixes this issue as that is the arn of AWS's Log Delivery Service account.  I have tested and verified this is working and allows logs to be written and blocks both console and sdk public object attempts