Add 'value_regex' to filters
This adds a new option to filters, that allows filtering based on
the output of a regex capture. This is useful for parsing data
that has a weird format.

Instead of comparing the contents of the 'resource value' (e.g. the
contents of a tag on a resource) with the 'value', it will instead
apply the regex to contents of the 'resource value', and compare
the result of the capture group defined in that regex with the 'value'.