Pull correct jquery and popper from CDNs
Bootstrap minified JS comes down with the CSS so we will continue
to use the local copy.  Rather than pull in and maintain the
jquery and popper JS files we need, convert to loading those from
their respective CDNs per bootstrap documentation
https://getbootstrap.com/docs/4.3/getting-started/introduction/#js
which also has the additional side benefit that we now have the
correct jquery and popper versions loading.