encode report data as bytes when the --raw output param is used
resolves custodian --raw throwing error #5269

I tried to get the encoding to work within json.dump but was getting another TypeError  'descriptor 'encode' requires a 'str' object but received a 'bytes'' so instead, I decided to encode it in memory then write to the filehandler. I considered using ```sys.version_info >= 3``` so it would still be done the old way in 2.7 but in testing, I found that this solution works in both py27 and py37 so keeping it the same for consistency makes sense. What do you think?