[metrics] push start time back to the top of the last hour
When calculating a start time for the metrics filter, zero out the
minutes so we always start from the top of an hour.

CloudWatch rounds down to the last 1-minute, 5-minute or 1-hour interval
based on its own retention schedule. By always rounding
down to the hour on our end, we accept 1 hour granularity but avoid
the burden of keeping pace with CloudWatch's retention details.

If CloudWatch rounds farther back than we do, we sometimes get
multiple data points back which leads to false positive matches.