tags - define replace behavior for normalize-tag
Provide an implementation for the `replace` sub-action of
`normalize-tag`. Replace tag values with a given value if the existing
tag matches a regex pattern.

Have NormalizeTag derive from UniversalTag, add tests

Change the superclass for NormalizeTag from Action to UniversalTag. This
brings a couple benefits:

* Broadens resource support for the `normalize-tag` action (from 24 to 88
resources as of today).

* Avoids duplicating logic such as fetching the appropriate service
client and farming work out to threads.