core - reduce filter - fix value_regex validation
Some logic from `ValueFilter` was extracted into `BaseValueFilter`
during the initial implementation of the `reduce` filter in #5874. That
did not include the `_validate_value_regex` method. This prevents reduce
filters with a `value_regex` parameter from validating. It does not
block _processing_ though, so the existing tests were passing without
error.

To address this:

* Add a validation step to a filter test to demonstrate the problem.
* Move `_validate_value_regex` from `ValueFilter` to `BaseValueFilter`.