docs - switch event id for AuthorizeSecurityGroup examples
Following the discussion in #6920, pull a security group ID from the
response rather than the request. This allows the event pattern to pick
up new authorizations made by group name rather than ID, for security
groups in the default VPC.

Note that this change would not work for the RevokeSecurityGroup events
- handling that case would require some magic under the hood to look for
and explicitly handle a groupName request parameter.