Use committed terraform state as pytest-terraform fixture
The new functionality and tests for the elasticsearch cross-cluster
search filter came with some saved Terraform bits that the tests weren't
using. One option is to remove that, but another is to convert the state
into pytest-terraform format and use it as a fixture. This is a sample
of the latter approach.