refactor: use LooseVersion instead of pkg_resources
Both distutils and pkg_resources have some warnings that it's bettter
not to use them.

It was decided to use LooseVersion beacuse LooseVersion is already
used in c7n (e.g., rds.py, core.py) so that it will be able to
handle its deprecation along with the rest of uses.