fix: get subnet_ids from all interfaces on an EC2 instance
An EC2 instance may have multiple network interfaces where each
interface may be deployed in different subnets.
As a result, EC2 instance is available in multiple subnets so
that we have to check all of them in 'subnet' filter, not only
default one

Use Case:

Need to make sure that EC2 instances is not deployed
in public subnets. In that case, we have to check both
primary and secondary interfaces because some of them may be
deployed in a public subnet.