aws - metrics-filter - s3 : fetch metrics from the bucket region
while we seem to be able to query the S3 API from multiple region endpoints (and have
quite the same results(*)), we can only query S3 CloudWatch Metrics from
the real bucket region (`LocationConstraint`).

this patch allows the `metrics` filter to query the real bucket region whenever
it's applied on a `s3` resource.

(*) One of the differences I'm aware of is that the `CreationDate` has not the same
meaning in all regions for bucket created outside of `us-east-1`:
* in `us-east-1` it's (as expected) the date of the creation of the bucket
* in other regions it's the date of the last synchronization of the bucket properties
between the region and `us-east-1` (after an update of the bucket).
Cf https://github.com/aws/aws-cli/issues/3597