aws - metrics-filter - automatically retrieve all related dimensions
In some cases (e.g. `s3`) you can have a metric split over multiple dimensions.

In the case of `s3`, the interesting metric `BucketSizeBytes` is detailed over
multiple `StorageType` dimensions (cf [documentation](https://docs.aws.amazon.com/AmazonS3/latest/userguide/metrics-dimensions.html#s3-cloudwatch-metrics))

To query this metric and make a policy based on the size taken by the objects, you either need to:
* restrict yourself to a subset of the storage type dimensions (e.g. `StandardStorage`)
* know in advance all the storage types used on your buckets (and hope they don't change)
* construct a complex metric filter with all the 22 (at this time) types as dimensions.

This patch introduce a new boolean argument to the `metrics` filter : `retrieve-all-dimensions` whose aim
is to simplify this query.

When the flag is set to `true`:
* c7n queries the CW Metrics API (`list_metrics()`) to list all the extra dimensions available
* then it builds a query for those dimensions, plus an extra query for the `SUM()` of all these metrics
* sends this query to CW Metrics API (using `get_metric_data()`)
* and makes the answer fits the existing format (the metric comparison being on the `SUM()` - and you
have all the extra dimensions available as metrics in `c7n.metrics`)
(NB: You loose the `Unit` which does not seem to be available with the `get_metric_data()` API)

As a consequence, you can now write a filter like:
```yaml
- name: s3_more_128Mb
  resource: aws.s3
  filters:
    - type: metrics
      name: BucketSizeBytes
      statistics: Average
      retrieve-all-dimensions: true
      days: 2
      period: 86400
      value: 134217728
      missing-value: 0
      op: gte
```
which will list all s3 bucket using more that 128Mb of space.

the detail of the metrics is available in `c7n.metrics`:
```json
    "c7n.metrics": {
      "AWS/S3.BucketSizeBytes.Average.2.GlacierS3ObjectOverhead": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 69591040.0
        }
      ],
      "AWS/S3.BucketSizeBytes.Average.2.GlacierStorage": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 24516200.0
        }
      ],
      "AWS/S3.BucketSizeBytes.Average.2.GlacierObjectOverhead": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 278364160.0
        }
      ],
      "AWS/S3.BucketSizeBytes.Average.2.StandardIAStorage": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 657660.0
        }
      ],
      "AWS/S3.BucketSizeBytes.Average.2.StandardStorage": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 35276207.0
        }
      ],
      "AWS/S3.BucketSizeBytes.Average.2": [
        {
          "Timestamp": "2023-09-16T13:33:00+00:00",
          "Average": 408405267.0
        }
      ]
    }
```

the `SUM()` of the metric accross all dimensions is in `AWS/S3.BucketSizeBytes.Average.2`, which is used for
the value comparison.
And all the dimensions are available as `AWS/S3.BucketSizeBytes.Average.2.XXXX` where XXXX is the extra dimension (
not available for value comparison)

Maybe related to #6151