Fix: Handles InvalidParameterValueException for refreshing trusted advisor checks (#9001)
As it was reported in #9001, we are also experiencing the application crashing with the error:

```
botocore.exceptions.ClientError: An error occurred (InvalidParameterValueException) when calling the RefreshTrustedAdvisorCheck operation: [UNREFRESHABLE_CHECK_ID_ERROR] Check c1dfprch07 is not refreshable.
```

Since I understand this comes from a change in AWS' APIs and until there's a better way to manage this, I'm proposing that we handle this exception at least so custodian does not crash.

aws - arn parse explicit value error on invalid (#9071)