Add a large c7n-left test and docs for jmespath
I thought there was a bug in c7n-left so I authored up a test which then
made me realize I had a bug in my filter.  I'm sending this PR just
in case we think having a larger test for c7n-left is useful even
though the functionality I was trying to verify is working as expected.