azure - wait for async tag operation results from tests
This attempts to address the flaky Azure test issues raised in #10152.
I've been able to reproduce the failures, but only in
certain situations.

This change does no harm to commits that were already consistently
working, but seems to make the tests reliably pass in cases where I was
seeing intermittent failures.